refactor($main): refactor application menu
- rename processCommunicator with ipcSender
- separate menu template from createMenu
