+++ src/constants/dialog-types.js
@@ -1,1 +1,2 @@
 export const MIDI_EXPORT = 'MIDI_EXPORT';
+export const INSPECT = 'INSPECT';
