+++ src/constants/event-types.js
@@ -1,0 +1,2 @@
+export const CONTEXT_MENU = 'contextmenu';
+export const DIALOG = 'dialog';
