+++ src/main/modules/createMenu.js
@@ -1,54 +1,6 @@
-import { app, Menu } from 'electron';
-import proccessCommunicator from './proccessCommunicator';
-import { MIDI_EXPORT } from '../../constants/message-types';
+import { Menu } from 'electron';
+import template from '../templates/menu';
 
 export default () => {
-  const application = {
-    label: 'Application',
-    submenu: [
-      {
-        role: 'about'
-      },
-      {
-        type: 'separator'
-      },
-      {
-        label: 'Quit',
-        click: () => {
-          app.quit();
-        }
-      }
-    ]
-  };
-
-  const edit = {
-    label: 'Edit',
-    submenu: [
-      {
-        label: 'Export MIDI',
-        click: () => {
-          proccessCommunicator.send(MIDI_EXPORT);
-        }
-      },
-      {
-        type: 'separator'
-      },
-      {
-        role: 'cut'
-      },
-      {
-        role: 'copy'
-      },
-      {
-        role: 'paste'
-      },
-      {
-        role: 'selectAll'
-      }
-    ]
-  };
-
-  const template = [application, edit];
-
   Menu.setApplicationMenu(Menu.buildFromTemplate(template));
 };
