+++ src/main/modules/createWindow.js
@@ -1,5 +1,5 @@
 import { BrowserWindow, screen } from 'electron';
-import proccessCommunicator from './proccessCommunicator';
+import ipcSender from './IpcSender';
 
 const winURL = process.env.NODE_ENV === 'development'
   ? `http://localhost:9080`
@@ -17,6 +17,6 @@ export default () => {
     titleBarStyle: 'hidden'
   });
   mainWindow.loadURL(winURL);
-  proccessCommunicator.appendWindow(mainWindow);
+  ipcSender.appendWindow(mainWindow);
   return mainWindow;
 };
